refactor #25: feature/place
